requirements.txt
flask
pandas
sentence-transformers
chromadb
openai
python-dotenv